requirements.txt
apache-airflow-providers-celery>=3.0.0
apache-airflow-providers-http>=4.0.0
python-dotenv>=1.0.0
